+++ src/graphics/left_bar.ts
