+++ src/graphics/right_bar.ts
